restructure sidebar components
